docs(did-comm): adapt inline docs for IDIDComm
